Techin510-Final-main/requirements.txt
streamlit
boto3
python-dotenv
pandas
google-generativeai